Ajustando a modelagem de dados

+++ z-tech-conta-bancodigital/src/main/java/br/com/ztech/repository/TipoTransacaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ztech.domain.TipoTransacao;
 
-public interface TipoTransacaoRepository extends JpaRepository<TipoTransacao, Long> {
+public interface TipoTransacaoRepository extends JpaRepository<TipoTransacao, Integer> {
 	
 }
